% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BF_power_analysis.R
\name{BF_Power_analysis}
\alias{BF_Power_analysis}
\title{Bayes Factor Power Analysis}
\usage{
BF_power_analysis(pars, Model)
}
\arguments{
\item{pars}{list(N, beta_M, beta_Y, sigma_M, sigma_Y, R, burnin, reps) for "Simple", list(N, beta_M, beta_Y, ssq_m_star, lambda, Sigma_Y, M_ind, Mcut, cutoff_M, R, burnin, reps) for "MCat", list(N, beta_M, beta_Y, ssq_y_star, tau, sigma_M, Y_ind, Ycut, cutoff_Y, R, burnin, reps)  for "YCat", and list(N, beta_M, beta_Y, ssq_m_star, ssq_y_star, lambda, Sigma_Y, M_ind, Mcut, cutoff_M, R, burnin, reps)  for "MYCat"}

\item{Model}{can be either "Simple", "MCat", "YCat", "MYCat".}
}
\value{
\subsection{The function returns the Bayes factor distribution obtained from "reps" datasets generated with the input parameter values}{
}
}
\description{
Bayes Factor Power Analysis
}
\details{
\subsection{Model}{

For further information about the models, see
\itemize{
\item \link[BFMediate]{PartialMed} for "Simple"
\item \link[BFMediate]{MeasurementMCat} for "MCat"
\item \link[BFMediate]{MeasurementYCat} for "YCat"
\item \link[BFMediate]{MeasurementMYCat} for "MYCat"
}
}
}
\examples{
pars = NULL
pars$M_ind = 2
pars$Y_ind = 2
pars$Mcut = pars$Ycut = 8
pars$N = 50
pars$beta_M = c(.5,1)
pars$beta_Y = c(.7, 1.5, 0)
pars$ssq_m_star = c(.5,.3)
pars$lambda = c(0,-.5) #the intercepts for the latent M indicators w. measurement error
pars$ssq_y_star = c(.2,.2)
pars$tau = c(0,-.5)
pars$cutoff_M = matrix(c(-100, 0, 1.6, 2, 2.2, 3.3, 6,  100,
                      -100, 0, 1, 2, 3, 4, 5, 100) ,ncol= pars$Mcut, byrow = TRUE)
pars$cutoff_Y =  matrix(c(-100, 0, 1.6, 2, 2.2, 3.3, 6,  100,
                         -100, 0, 1, 2, 3, 4, 5, 100) ,ncol= pars$Ycut, byrow = TRUE)
pars$R = 10000
pars$burnin = 2000
pars$reps = 100  

BF_Power_analysis(pars = pars, model = "MYCat")    # The analysis takes several minutes
}
