% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFMediate-package.R
\docType{package}
\name{BFMediate-package}
\alias{BFMediate}
\alias{BFMediate-package}
\title{BFMediate: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
