% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFmediate_Rfunction.R
\name{BFMediate}
\alias{BFMediate}
\title{Title}
\usage{
BFMediate(Data, Model, Prior, R, burnin)
}
\arguments{
\item{Data}{sdf}

\item{Model}{sdf}

\item{Prior}{sdf}

\item{R}{sdf}

\item{burnin}{sdf}
}
\value{
sdf
}
\description{
Title
}
