% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenMoments.R
\name{GenMoments}
\alias{GenMoments}
\title{Title}
\usage{
GenMoments(Data, post, pars, R)
}
\arguments{
\item{Data}{asdf}

\item{post}{asdf}

\item{pars}{asdf}

\item{R}{asdf}
}
\value{
asdf
}
\description{
Title
}
