% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measurement_Cont.R
\name{Measurement_Cont}
\alias{Measurement_Cont}
\title{Measurement_Cont}
\usage{
Measurement_Cont(Data, BF)
}
\arguments{
\item{Data}{asd}

\item{Prior}{asd}

\item{Mcmc}{asd}
}
\value{
asd
}
\description{
Measurement_Cont
}
