% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementCont.R
\name{MeasurementCont}
\alias{MeasurementCont}
\title{Sampler for Partial Mediation Model with Multiple Continuous Indicator for the Mediator and/or DV}
\usage{
MeasurementCont(Data, Prior, R, burnin)
}
\arguments{
\item{Data}{list(X, m_tilde, y_tilde)}

\item{Prior}{list(A_M,A_Y)}

\item{R}{number of MCMC iterations, default = 10000}
}
\value{
\describe{
\item{beta_M(R X 2)}{matrix of eq.1 coefficients' draws}
\item{beta_Y(R X 3)}{matrix of eq.2 coefficients' draws}
\item{lambda (M_ind X 2 X R)}{array of mediator indicator coefficients' draws. Each slice is one draw, where rows represent the indicator equation and columns are the coefficients. The ntercept and the slope of the first equation are fixed to 0 and 1 respectively.}
\item{tau (Y_ind X 2 X R)}{array of dependent variable indicator coefficients' draws. Each slice is one draw, where rows represent the indicator equation and columns are the coefficients. The ntercept and the slope of the first equation are fixed to 0 and 1 respectively.}
\item{ssq_m_star(R X M_ind)}{Matrix of mediator indicator equations' coefficients' error variance draws}
\item{ssq_y_star(R X Y_ind)}{Matrix of dependent variable indicator equations' coefficients' error variance draws}
\item{mu_draw}{vector of means indexing MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
\item{var_draw}{vector of variance indexing MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
}
}
\description{
Estimates a partial mediation model with multiple continuous indicators for the mediator and the dependent variable using
Stan's Hamiltonian Monte Carlo (HMC)
}
\details{
\subsection{Model}{

\eqn{M = \beta_{0M} + X\beta_1 + U_M}{M = \beta_0M + X\beta_1 + U_M} (eq.1) \cr
\eqn{Y = \beta_{0Y} + M\beta_2 + X\beta_3 + U_Y}{Y = \beta_0Y + M\beta_2 + X\beta_3 + U_Y} (eq.2) \cr
}

\subsection{Indicator equations}{

\eqn{m^*_1 = M + U_{m^*_1}}{m*_1 = M + U_{m*_1}}
\eqn{m^*_2 = \lambda_{20} + M\lambda_{21} + U_{m^*_2}}{m*_2 = \lambda_20 + M\lambda_21 + U_{m*_2}}
...
\eqn{m^*_K = \lambda_{K0} + M\lambda_{K1} + U_{m^*_K}}{m*_K = \lambda_K0 + M\lambda_K1 + U_{m*_K}}
\eqn{y^*_1 = M + U_{y^*_1}}{y*_1 = M + U_y*_1}
\eqn{y^*_2 = \tau_{20} + M\tau_{21} + U_{y^*_2}}{y*_2 = \tau_20 + M\tau_21 + U_{y*_2}}
...
\eqn{y^*_L = \tau_{L0} + M\tau_{L1} + U_{y^*_L}}{y*_L = \tau_L0 + M\tau_L1 + U_{y*_L}}
}

\subsection{Prior specification:}{

\eqn{\beta_{0M}} \eqn{\sim}{~} \eqn{N(0,100)}, \eqn{\beta_{0Y}} \eqn{\sim}{~} \eqn{N(0,100)}  \cr
\eqn{\beta_1} \eqn{\sim}{~} \eqn{N(0,100)}, \eqn{\beta_2} \eqn{\sim}{~} \eqn{N(0,100), \beta_3} \eqn{\sim}{~} \eqn{N(0,1)}  \cr
\eqn{\lambda_{20},...,\lambda_{K0}} \eqn{\sim}{~}  \eqn{N(0,100)}  \cr
\eqn{\sigma^2_{m*_1}, ..., \sigma^2_{m*_K}} \eqn{\sim}{~} \eqn{(nu*S)/\chi^2_{nu}} \cr
\eqn{\tau_{20}, ..., \tau_{L0}} \eqn{\sim}{~}  \eqn{N(0,100)} \cr
\eqn{\sigma^2_{y*_1}, ..., \sigma^2_{y*_L}} \eqn{\sim}{~} \eqn{(nu*S)/\chi^2_{nu}} \cr
}

\subsection{Argument Details}{
}

\subsection{\code{Data = list(X, m_tilde, y_tilde)}}{

\describe{
\item{X(N x 1)}{treatment variable vector}
\item{m_tilde(N x M_ind)}{ediator indicators' matrix }
\item{y_tilde(N x Y_ind)}{dependent variable indicators' matrix}
}
}

\subsection{\code{Prior = list(A_M,A_Y)} \emph{[optional]}}{

\describe{
\item{A_M}{vector of coefficients' prior variances of eq.1, default = rep(100,2)}
\item{A_Y}{vector of coefficients' prior variances of eq.2, default = c(100,100,1)}
}
}
}
\examples{
library(rstan)
SimMeasurementCont = function( beta_M, beta_Y , lambda, tau, m_ind, y_ind, sigma_M, sigma_m_star,
                              sigma_y, sigma_y_star, N, X) {

  m_star = matrix(double(N*m_ind),ncol = m_ind); y_star = matrix(double(N*y_ind),ncol = y_ind)
  eps_m_star = matrix(double(N*m_ind),ncol = m_ind); eps_y_star = matrix(double(N*y_ind),ncol = y_ind)
  eps_M = rnorm(N)*sigma_M # generate errors for M (independent)
  eps_Y = rnorm(N)*sigma_y # generate errors for y (independent)
  M = beta_M[1] + X*beta_M[2]  + eps_M # generate latent mediator M
  y = beta_Y[1] +  M*beta_Y[2] + X*beta_Y[3] + eps_Y # generate dependent variable

  eps_m_star[,1]=rnorm(N)*sigma_m_star[1] # generate errors for m_star (independent)
  m_star[,1] =  M + eps_m_star[,1] # generate observed mediator indicators m_star
  if(m_ind>1){
    for(i in 2:(m_ind))   {
      eps_m_star[,i]=rnorm(N)*sigma_m_star[i] # generate errors for m_star (independent)
      m_star[,i] =  lambda[(i-1),1] + M*lambda[(i-1),2] + X*beta_Y[-c(1,2)] + eps_m_star[,i]
    }
  }

  eps_y_star[,1]=rnorm(N)*sigma_y_star[1] # generate errors for y_star (independent)
  y_star[,1] =  y + eps_y_star[,1] # generate observed dependent variable indicators y_star
  if(y_ind>1){
    for(i in 2:(y_ind)){
      eps_y_star[,i]=rnorm(N)*sigma_y_star[i] # generate errors for y_star (independent)
      y_star[,i] =  tau[(i-1),1] + y*tau[(i-1),2] + eps_y_star[,i]
    }
  }
  list(X = X, M = M, m_star = m_star, y = y, y_star=y_star)
}
m_ind = 2; y_ind = 2;
sigma_M = 1^.5 # error std M
sigma_y = 1^.5 # error std y
sigma_m_star = c(.3,.5)^.5 #c(1,2)^.5
sigma_y_star = c(.5,.3)^.5  #c(2,1)^.5
beta_M = c(.5,1)
beta_Y = c(1,3,0)
lambda = matrix(c(1,1.5),ncol=2)
tau = matrix(c(1,2),ncol = 2)
k=length(beta_M)-1
nobs = 1000   # number of observations
X = runif(nobs) # generate random X from a uniform distribution
Data = SimMeasurementCont( beta_M, beta_Y , lambda, tau, m_ind, y_ind, sigma_M, sigma_m_star,
                          sigma_y, sigma_y_star, nobs, X)
R = 5000; burnin = 1000

A_M=rep(100,2);
A_Y=c(100,100,1)

#Estimation
out = MeasurementCont(Data = Data, Prior = list(A_M = A_M, A_Y = A_Y),R=5000, burnin = 1000)

#Results
colMeans(out$beta_M)
colMeans(out$beta_Y)

BFMeasurementCont = exp(BFSD(Post = out , Prior = A_Y[3],burnin = 0))
}
