% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementMCat.R
\name{MeasurementMCat}
\alias{MeasurementMCat}
\title{asd}
\usage{
MeasurementMCat(Data, Prior, R = 10000)
}
\arguments{
\item{Data}{asd}

\item{Prior}{asd}

\item{Mcmc}{asd}
}
\value{
asd
}
\description{
asd
}
