% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementYCat.R
\name{MeasurementYCat}
\alias{MeasurementYCat}
\title{qwerqw}
\usage{
MeasurementYCat(Data, Prior, Mcmc)
}
\arguments{
\item{Data}{qwerqwer}

\item{Prior}{qwerwer}

\item{Mcmc}{qwererrr}
}
\value{

}
\description{
qwerqw
}
