% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mediation_Ordered_Multi_Merr.R
\name{Mediation_Ordered_Multi_Merr}
\alias{Mediation_Ordered_Multi_Merr}
\title{asdfsaf}
\usage{
Mediation_Ordered_Multi_Merr(Data, Prior, Mcmc)
}
\arguments{
\item{Data}{asdf}

\item{Prior}{asdf}

\item{Mcmc}{aasdf}
}
\value{

}
\description{
asdfsaf
}
