% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialMed.R
\name{PartialMed}
\alias{PartialMed}
\title{Gibbs Sampler for Partial Mediation Model}
\usage{
PartialMed(Data, pars, R)
}
\arguments{
\item{Data}{list(X, M, Y)}

\item{pars}{list(A_M,A_Y)}

\item{R}{number of MCMC iterations, default = 10000}
}
\value{
a list containing
\tabular{ll}{
\code{beta_1(R X 2) } \tab  matrix of eq.1 coefficients' posterior draws \cr
\code{beta_2(R X 3) } \tab  matrix of eq.2 coefficients' posterior draws \cr
\code{ssq_M(R X 1) } \tab  vector of eq.1 error variance posterior draws \cr
\code{ssq_Y(R X 1) } \tab  vector of eq.2 error variance posterior draws \cr
\code{mu_draw } \tab  vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor) \cr
\code{var_draw } \tab  vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor) \cr
}
}
\description{
Estimates a partial mediation model using series of Gibbs Samplers
}
\details{
\emph{Argument Details}

\code{Data = list(X, M, Y)}

\tabular{ll}{
\code{X(N x 1) } \tab treatment variable vector \cr
\code{M(N x 1) } \tab mediator vector \cr
\code{Y(N x 1) } \tab dependent variable vector \cr
}

\code{pars = list(A_M,A_Y)} \emph{\link{optional}}

\tabular{ll}{
\code{A_M }   \tab vector of coefficients' prior variances of eq.1, default = rep(100,2) \cr
\code{A_Y }   \tab vector of coefficients' prior variances of eq.2, default = c(100,100,1) \cr
}
}
\examples{
simPartialMed = function(beta_1,beta_2, sigma_M, sigma_Y,N,X) {
  eps_M = rnorm(N)*sigma_M # generate errors for M (independent)
  eps_Y = rnorm(N)*sigma_Y # generate errors for Y (independent)
  M = beta_1[1] + beta_1[2] * X + eps_M # generate latent mediator M
  Y = beta_2[1] + beta_2[2] * M + beta_2[3] * X + eps_Y # generate dependent variable
  list(X = X, M = M, Y = Y)
}

# Set up data generating parameters
N = 1000 # number of observations
sigma_M = .2^.5 # error std M
sigma_Y = .2^.5 # error std Y
beta_1 = c(1, .3) # beta_0M and beta_1
beta_2 = c(1, .5, .01) # beta_0Y, beta_2, beta_3
X = rnorm(N,mean = 1,sd = 1)# generate random X
# Generate data based on parameters
Data = simPartialMed(beta_1,beta_2,sigma_M,sigma_Y,N,X)

#Estimation
A_M = c(100,100); #Prior variance for beta_0M, beta_1
A_Y = c(100,100,1) #Prior variance for beta_0Y, beta_2, beta_3
R = 2000
out = PartialMed(Data=Data, pars = list(A_M=A_M, A_Y=A_Y), R = R)

#Computing Bayes factor
BFPartialMed = exp(BFSD(post = out , prior = A_Y[3], burnin = 0))
}
