% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rordprobitGibbs_me_M_multi_merr.R
\name{rordprobitGibbs_me_M_multi_merr_cpp}
\alias{rordprobitGibbs_me_M_multi_merr_cpp}
\title{asd}
\usage{
rordprobitGibbs_me_M_multi_merr_cpp(Data, Prior, Mcmc)
}
\arguments{
\item{Data}{asd}

\item{Prior}{asd}

\item{Mcmc}{asd}
}
\value{

}
\description{
asd
}
