% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rordprobitGibbs_me_multi_merr.R
\name{rordprobitGibbs_me_multi_merr_cpp}
\alias{rordprobitGibbs_me_multi_merr_cpp}
\title{qwerqw}
\usage{
rordprobitGibbs_me_multi_merr_cpp(Data, Prior, Mcmc)
}
\arguments{
\item{Data}{qwerqwer}

\item{Prior}{qwerwer}

\item{Mcmc}{qwererrr}
}
\value{

}
\description{
qwerqw
}
