% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFMediateApp.R
\name{run_BFMediate}
\alias{run_BFMediate}
\title{Run BFMediateApp}
\usage{
run_BFMediate()
}
\description{
Launches a Shiny app that shows a demo of what can be done with
the \code{BFMediate} package.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  run_BFMediate()
}
}
