% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFMediate-package.R
\docType{package}
\name{BFMediate-package}
\alias{BFMediate}
\alias{BFMediate-package}
\title{BFMediate: Mediation Analysis Using a Combination of Bayesian Estimation Methods, Latent Variable Models, and Bayes Factors}
\description{
The focus is on measuring evidence for (full) mediation using Bayes factors. The package covers models with measurement error and discretization in the mediator (M) and/or the dependent variable. For further details, see the paper, Measuring Evidence for Mediation in the Presence of Measurement Error by Laghaie and Otter (2020).
}
\keyword{internal}
