% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSD.R
\name{BFSD}
\alias{BFSD}
\title{BFSD}
\usage{
\code{BFSD(post,prior,burnin)}
}
\arguments{
\item{post}{output from PartialMed or any of measurement models}

\item{prior}{prior variance of the direct effect}

\item{burnin}{number of MCMC draws before the posterior is converged, default = R/5}
}
\value{
log(BF_01), which is the evidence in favor of the full mediation model (see Laghaie and Otter (2020) for guidelines on how to interpret BF_01)
}
\description{
Computes Bayes factors for the partial mediation model using Savage-Dickey approximation
}
