% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSD.R
\name{BFSD}
\alias{BFSD}
\title{Bayes factor for full mediation}
\usage{
BFSD(post,prior,burnin)
}
\arguments{
\item{post}{output from PartialMed or any of measurement models}

\item{prior}{prior variance of the direct effect}

\item{burnin}{number of MCMC draws before the posterior is converged, default = R/5}
}
\value{
log(BF_01), which is the evidence in favor of the full mediation model (see Laghaie and Otter (2020) for guidelines on how to interpret BF_01)
}
\description{
Computes Bayes factors for the partial mediation model using Savage-Dickey approximation
}
\examples{
simPartialMed = function(beta_1,beta_2, sigma_M, sigma_Y,N,X) {
  eps_M = rnorm(N)*sigma_M # generate errors for M (independent)
  eps_Y = rnorm(N)*sigma_Y # generate errors for Y (independent)
  M = beta_1[1] + beta_1[2] * X + eps_M # generate latent mediator M
  Y = beta_2[1] + beta_2[2] * M + beta_2[3] * X + eps_Y # generate dependent variable
  list(X = X, M = M, Y = Y)
}

# Set up data generating parameters
N = 1000 # number of observations
sigma_M = .2^.5 # error std M
sigma_Y = .2^.5 # error std Y
beta_1 = c(1, .3) # beta_0M and beta_1
beta_2 = c(1, .5, 0) # beta_0Y, beta_2, beta_3
X = rnorm(N,mean = 1,sd = 1)# generate random X
# Generate data based on parameters
Data = simPartialMed(beta_1,beta_2,sigma_M,sigma_Y,N,X)

#Estimation
A_M = c(100,100); #Prior variance for beta_0M, beta_1
A_Y = c(100,100,1) #Prior variance for beta_0Y, beta_2, beta_3
R = 2000
out = PartialMed(Data=Data, pars = list(A_M=A_M, A_Y=A_Y), R = R)
#Computing Bayes factor
BFPartialMed = exp(BFSD(post = out , prior = A_Y[3], burnin = R/5))
}
\seealso{
For simulating data from simple mediation model see \link[BFMediate]{PartialMed}
}
