% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFSD.R
\name{BFSD}
\alias{BFSD}
\title{BFSD}
\usage{
BFSD(Data, post, prior, model, burnin, M)
}
\arguments{
\item{Data}{Data}

\item{post}{Posterior}

\item{prior}{Prior}

\item{model}{Model}

\item{burnin}{Burnin}

\item{M}{MMMMM}
}
\value{
HAHAHAHA
}
\description{
BFSD
}
