% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementCont.R
\name{MeasurementCont}
\alias{MeasurementCont}
\title{MeasurementCont}
\usage{
MeasurementCont(Data, Prior, R)
}
\arguments{
\item{Data}{asd}

\item{Prior}{asd}

\item{Mcmc}{asd}
}
\value{
asd
}
\description{
MeasurementCont
}
