% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementMCat.R
\name{MeasurementMCat}
\alias{MeasurementMCat}
\title{Sampler for Partial Mediation Model with Multiple Categorical Indicator for the Mediator}
\usage{
MeasurementMCat(Data,Prior,R)
}
\arguments{
\item{Data}{list(X, m_star, Y)}

\item{Prior}{list(A_M,A_Y)}

\item{R}{number of MCMC iterations, default = 10000}
}
\value{
\describe{
\item{beta_1(R X 2)}{matrix of eq.1 coefficients' draws}
\item{beta_2(R X 3)}{matrix of eq.2 coefficients' draws}
\item{lambda (M_ind X 2 X R)}{array of indicator coefficients' draws. Each slice is one draw, where rows represent the indicator equation and columns are the coefficients. All Slope coefficients as well as intercept of the first equation are fixed to 1 and 0 respectively.}
\item{ssq_m_star(R X M_ind)}{Matrix of indicator equations' coefficients' error variance draws}
\item{ssq_Y(R X 1)}{vector of eq.2 error variance draws}
\item{mu_draw}{vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
\item{var_draw}{vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
}
}
\description{
Estimates a partial mediation model with multiple categorical indicator for the mediator and observed dependent variable using a mixture of Metropolis-Hastings and Gibbs sampling
}
\details{
\subsection{Model}{

\tabular{ll}{
M = beta_0M + Xbeta_1 + U_M  \tab [eq.1]\ \cr
Y = beta_0Y + Mbeta_2 + Xbeta_3 + U_Y \tab [eq.2] \cr
}

Indicator equations:

\tabular{lcl}{
m*_1   \tab = \tab M + U_m*_1 \cr
˜m_1 \tab = \tab  OrdProbit(m*_1,C_m_1) \cr
m*_2   \tab = \tab lambda_01 + M + U_m*_2 \cr
˜m_2   \tab = \tab OrdProbit(m*_2,C_m_2) \cr
... \tab  \tab  \cr
m*_k  \tab =  \tab lambda_0k-1 + M + U_m*_k \cr
˜m_k \tab = \tab  OrdProbit(m*_k,C_m_k) \cr
}
}

\subsection{Argument Details}{
}

\subsection{\code{Data = list(X, m_star, Y)}}{

\describe{
\item{X(N x 1)}{treatment variable vector}
\item{m_star(N x M_ind)}{mediator indicators' matrix}
\item{Y(N x 1)}{dependent variable vector}
}
}

\subsection{\code{Prior = list(A_M,A_Y)} [optional]}{

\describe{
\item{A_M}{vector of coefficients' prior variances of eq.1, default = rep(100,2)}
\item{A_Y}{vector of coefficients' prior variances of eq.2, default = c(100,100,1)}
}
}
}
\examples{
SimMeasurementMCat = function(X, beta_1, cutoff_M, beta_2, Sigma_Y, M_ind, beta_m_tilde, ssq_m_tilde){

nobs = dim(X)[1]
m_star = m_tilde = matrix(double(nobs*M_ind), ncol = M_ind)

M = beta_1[1] + beta_1[2] * X + rnorm(nobs)  #cbind(rep(1,nobs),X)\%*\%beta_1 + rnorm(nobs)

for(i in 1: M_ind){
  m_tilde[,i] = beta_m_tilde[i] + M + sqrt(ssq_m_tilde[i])*rnorm(nobs);
  m_star[,i] = cut(m_tilde[,i], br = cutoff_M[i,], right=TRUE, include.lowest = TRUE, labels = FALSE)
}

Y = beta_2[1] + beta_2[2] * M + beta_2[3] * X + rnorm(nobs)*Sigma_Y
                            #cbind(rep(1,nobs),cbind(M,X))\%*\%beta_2 + rnorm(nobs)

return(list(Y = Y, M = M, m_star = m_star, X = X,
            beta_1 = beta_1, beta_2 = beta_2,
            beta_m_tilde = beta_m_tilde, ssq_m_tilde = ssq_m_tilde, m_tilde = m_tilde, cutoff_M = cutoff_M,
            k_M=dim(cutoff_M)[2]-1, M_ind=M_ind))
}

M_ind = 2
Mcut =  8
nobs= 500
X=as.matrix(runif(nobs,min=0, max=1))
beta_1 = c(.5,1)
beta_2 = c(1, 2, 0)
Sigma_Y = 1^.5
ssq_m_tilde = c(.5,.7)
beta_m_tilde = c(0,-.5)   #the intercepts for the latent M indicators w. measurement
                          #error (first intercept should always be 0)
cutoff_M = matrix(c(-100, 0, 1.6, 2, 2.2, 3.3, 6,  100,
                    -100, 0, 1, 2, 3, 4, 5, 100) ,ncol= Mcut, byrow = T)
DataMCat = SimMeasurementMCat(X, beta_1, cutoff_M, beta_2, Sigma_Y, M_ind, beta_m_tilde, ssq_m_tilde)

#estimation
Mcut = max(DataMCat$m_star) +1
Data = list(X=cbind(rep(1,length(DataMCat$X)),DataMCat$X), m_star=as.matrix(DataMCat$m_star),
                    Y= as.matrix(DataMCat$Y) ,k=Mcut-1, M_ind=dim(DataMCat$m_star)[2])
# Mcmc = list(R = 10000)
out = MeasurementMCat(Data=Data, R=R) #Mcmc=Mcmc)


#results
colMeans(out$beta_1)
colMeans(out$beta_2)
apply(out$cutoff_M,c(1,2),FUN = mean)
}
