% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementMYCat.R
\name{MeasurementMYCat}
\alias{MeasurementMYCat}
\title{asdfsaf}
\usage{
MeasurementMYCat(Data, Prior, Mcmc)
}
\arguments{
\item{Data}{asdf}

\item{Prior}{asdf}

\item{Mcmc}{aasdf}
}
\value{

}
\description{
asdfsaf
}
