% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementYCat.R
\name{MeasurementYCat}
\alias{MeasurementYCat}
\title{Estimates a partial mediation model with multiple categorical indicator for the dependent variable}
\usage{
\code{MeasurementYCat(Data,Prior,R)}
}
\arguments{
\item{Data}{list(X, M, y_star)}

\item{Prior}{list(A_M,A_Y)}

\item{R}{number of MCMC iterations, default = 10000}
}
\value{
\tabular{ll}{
\code{beta_1(R X 2) } \tab  matrix of eq.1 coefficients' draws \cr
\code{beta_2(R X 3) } \tab  matrix of eq.2 coefficients' draws \cr
\code{tau (Y_ind X 2 X R) } \tab array of indicator coefficients' draws. Each slice is one draw, where rows represent the indicator equation and columns are the coefficients. All Slope coefficients as well as intercept of the first equation are fixed to 1 and 0 respectively. \cr
\code{ssq_y_star(R X Y_ind) } \tab  Matrix of indicator equations' coefficients' error variance draws \cr
\code{ssq_M(R X 1) } \tab  vector of eq.1 error variance draws \cr
\code{mu_draw } \tab  vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor) \cr
\code{var_draw } \tab  vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor) \cr
}
}
\description{
Estimates a partial mediation model with multiple categorical indicator for the dependent variable
}
\details{
\subsection{Model}
\tabular{ll}{
M = beta_0M + Xbeta_1 + U_M  \tab \emph{[eq.1]} \cr
Y = beta_0Y + Mbeta_2 + Xbeta_3 + U_Y \tab \emph{[eq.2]} \cr
}
Indicator equations:
\tabular{lcl}{
y*_1   \tab = \tab M + U_y*_1 \cr
 ˜y_1  \tab = \tab  OrdProbit(y*_1,C_y_1) \cr
 y*_2    \tab = \tab tau_01 + M + U_y*_2 \cr
 ˜y_2  \tab = \tab OrdProbit(y*_2,C_y_2) \cr
 ... \tab  \tab  \cr
  y*_l   \tab =  \tab tau_0l-1 + M + U_y*_l \cr
 ˜y_l  \tab = \tab OrdProbit(y*_l,C_y_l) \cr
}


\subsection{Argument Details}
\code{Data = list(X, M, y_star)}
\tabular{ll}{
\code{X(N x 1) } \tab treatment variable vector \cr
\code{M(N x 1) } \tab mediator vector  \cr
\code{y_star(N x Y_ind) } \tab dependent variable indicators' matrix \cr
}

\code{Prior = list(A_M,A_Y)} \emph{[optional]}
\tabular{ll}{
\code{A_M }   \tab vector of coefficients' prior variances of eq.1, default = rep(100,2) \cr
\code{A_Y }   \tab vector of coefficients' prior variances of eq.2, default = c(100,100,1) \cr
}
}
