% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialMed.R
\name{PartialMed}
\alias{PartialMed}
\title{Partial Mediation}
\usage{
PartialMed(Data, pars, R = 10000)
}
\arguments{
\item{Data}{Data}

\item{pars}{Parameters}

\item{R}{Dunno what is it}
}
\value{
something
}
\description{
Partial Mediation
}
