% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialMed.R
\name{PartialMed}
\alias{PartialMed}
\title{Gibbs Sampler for Partial Mediation Model}
\usage{
PartialMed(Data, Pars, R)
}
\arguments{
\item{Data}{list(X, M, Y)}

\item{Pars}{list(A_M,A_Y)}

\item{R}{number of MCMC iterations, default = 10000}
}
\value{
a list containing
\describe{
\item{beta_1(R X 2)}{matrix of eq.1 coefficients' posterior draws}
\item{beta_2(R X 3)}{matrix of eq.2 coefficients' posterior draws}
\item{ssq_M(R X 1)}{vector of eq.1 error variance posterior draws}
\item{ssq_Y(R X 1)}{vector of eq.2 error variance posterior draws}
\item{mu_draw}{vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
\item{var_draw}{vector of means of MCMC draws of the direct effect (used in BFSD to compute Bayes factor)}
}
}
\description{
Estimates a partial mediation model using series of Gibbs Samplers
}
\details{
\subsection{Argument Details}{
}

\subsection{\code{Data = list(X, M, Y)}}{

\describe{
\item{X(N x 1) }{treatment variable vector}
\item{M(N x 1) }{mediator vector}
\item{Y(N x 1) }{dependent variable vector}
}
}

\subsection{\code{Pars = list(A_M,A_Y)} [optional]}{

\describe{
\item{A_M}{vector of coefficients' prior variances of eq.1, default = rep(100,2)}
\item{A_Y}{vector of coefficients' prior variances of eq.2, default = c(100,100,1)}
}
}
}
\examples{
simPartialMed = function(beta_1,beta_2, sigma_M, sigma_Y,N,X) {
  eps_M = rnorm(N)*sigma_M # generate errors for M (independent)
  eps_Y = rnorm(N)*sigma_Y # generate errors for Y (independent)
  M = beta_1[1] + beta_1[2] * X + eps_M # generate latent mediator M
  Y = beta_2[1] + beta_2[2] * M + beta_2[3] * X + eps_Y # generate dependent variable
  list(X = X, M = M, Y = Y)
}

# Set up data generating parameters
N = 1000 # number of observations
sigma_M = .2^.5 # error std M
sigma_Y = .2^.5 # error std Y
beta_1 = c(1, .3) # beta_0M and beta_1
beta_2 = c(1, .5, 0) # beta_0Y, beta_2, beta_3
X = rnorm(N,mean = 1,sd = 1)# generate random X
# Generate data based on parameters
Data = simPartialMed(beta_1,beta_2,sigma_M,sigma_Y,N,X)

#Estimation
A_M = c(100,100); #prior variance for beta_0M, beta_1
A_Y = c(100,100,1) #prior variance for beta_0Y, beta_2, beta_3
R = 2000
out = PartialMed(Data=Data, Pars = list(A_M=A_M, A_Y=A_Y), R = R)
}
